requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
faiss-cpu>=1.8.0
tavily-python>=0.5.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.9.0
sse-starlette>=2.1.0
httpx>=0.27.0
